backend/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0

# LangChain stack (versões compatíveis)
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=2.0.0

# File handling
pypdf>=4.0.0

# Utilities
python-multipart>=0.0.6
requests>=2.31.0
